airflow/requirements.txt
apache-airflow-providers-trino==4.3.1
psycopg2-binary==2.9.9
apache-airflow-providers-docker==3.9.0
astronomer-cosmos==1.6.0
dbt-core==1.7.0
dbt-trino==1.7.0
pydantic==2.5.0
python-dotenv==1.0.0
